refactoring of parent path and api url construction
