+++ pystackpath/stacks/deliverysites/delivery_domain.py
@@ -1,16 +1,14 @@
-from ..util import BaseObject, pagination_query, PageInfo
+from pystackpath.util import BaseObject, pagination_query, PageInfo
 
 
-class DeliverySites(BaseObject):
-    base_api = "/delivery"
-
+class DeliveryDomains(BaseObject):
     def index(self, first="", after="", filter="", sort_by=""):
         """
         Retrieve the delivery domains configured on a site
         :return: a list of domains configures on a site
         """
         pagination = pagination_query(first=first, after=after, filter=filter, sort_by=sort_by)
-        response = self._client.get("{}/v1/stacks/{}/delivery-domains".format(self.base_api, self._parent_id),
+        response = self._client.get(f"{self._base_api}/delivery-domains",
                                     params=pagination)
         response.raise_for_status()
         items = [self.loaddict(item) for item in response.json()["results"]]
